requirements.txt
fastapi==0.104.1
httpx==0.25.0
python-dotenv==1.0.0
python_Levenshtein==0.23.0
uvicorn==0.23.2
